refactor(runtime): split layout-manager.ts into transport/manager/adaptors
layout-manager.ts (727 LOC) mixed three concerns; split at the natural
seams with no behavior change:

- layout-transport.ts — GraphQLTransportWsDaemonTransport,
  DaemonTransport, DaemonEnvelope, TransportStatus, WebSocketLike,
  reconnect constants
- layout-adaptors.ts — ExperienceAdaptor, AdaptorHelpers, SlotErrorInfo,
  UnmountFn, SlotElementLike, htmlAdaptor, jsonAdaptor,
  moduleFederationAdaptor, loadRemoteContainer, MF bootstrap interfaces,
  defaultCreateSlotElement, defaultRenderSlotFallback, defaultAdaptors
- layout-manager.ts — LayoutManager, LayoutManagerConfig, ActiveSlot,
  LayoutHostLike (imports the other two)

SlotElementLike lives with the adaptors (they consume it) to avoid an
import cycle. layout-manager.ts keeps compatibility re-exports of every
name it previously exported, so existing imports — including the test
suites importing from '../layout-manager' — are untouched. index.ts now
re-exports from the owning modules (adds LayoutHostLike, previously
unexported). The ADR-056 boundary test scans the two new files so the
moved code stays inside the framework-neutral gate.

Each file is now under 360 LOC (336/133/299).

Refs #253, #254, #258. ADR-055, ADR-056, ADR-057.
